requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-qdrant
langchain-text-splitters
openai
qdrant-client
sentence-transformers
tiktoken
redis
numpy
scikit-learn
pypdf
rank_bm25
fastapi
uvicorn[standard]
pydantic
langchain-classic
